+++ src/core/logger.py
@@ -16,7 +16,7 @@ logger.setLevel(logging.DEBUG)
 
 
 def _linkrot_formatter(record: logging.LogRecord) -> str:
-    msg_date = datetime.fromtimestamp(record.created, tz=UTC).strftime("%B %d, %Y @ %I:%M:%S %p")
+    msg_date = datetime.fromtimestamp(record.created, tz=UTC).strftime("%B %d, %Y @ %I:%M:%S %p %Z")
     return f""":warning: Webring Alert :warning:
 Alert level: **{record.levelname.capitalize()}**
 Date: {msg_date}